[功能] 后端-做了转交操作的，都要在处理历史中留记录，做了转交操作的人，也可以在我的已办中查到这条工单。
#[1178993405952000]后端-做了转交操作的，都要在处理历史中留记录，做了转交操作的人，也可以在我的已办中查到这条工单。 http://192.168.0.96:8090/demo/rdm.html#/task-detail/939050947543040/939050947543050/1178993405952000

+++ src/main/java/neatlogic/framework/process/stephandler/core/ProcessStepHandlerBase.java
@@ -372,11 +372,29 @@ public abstract class ProcessStepHandlerBase implements IProcessStepHandler {
             throw new ProcessTaskStepNoMatchedWorkerException();
         }
 
-        ProcessTaskStepUserVo processTaskStepUserVo = new ProcessTaskStepUserVo();
-        processTaskStepUserVo.setProcessTaskStepId(currentProcessTaskStepVo.getId());
-        processTaskStepUserVo.setUserType(ProcessUserType.MAJOR.getValue());// 只删除主处理人人
-        processTaskCrossoverMapper.deleteProcessTaskStepUser(processTaskStepUserVo);
-
+        if (!oldUserList.isEmpty()) {
+            boolean changeMajorUser = true;
+            ProcessTaskStepUserVo oldUserVo = oldUserList.get(0);
+            for (ProcessTaskStepWorkerVo workerVo : workerSet) {
+                if (Objects.equals(workerVo.getUuid(), oldUserVo.getUserUuid())) {
+                    changeMajorUser = false;
+                    break;
+                }
+            }
+            if (changeMajorUser) {
+                oldUserVo.setUserType(ProcessTaskStepUserType.HISTORY_MAJOR.getValue());
+                oldUserVo.setStatus(ProcessTaskStepUserStatus.DONE.getValue());
+                int count = processTaskCrossoverMapper.updateProcessTaskStepMajorUserUserTypeAndStatus(oldUserVo);
+                System.out.println("count = " + count);
+                List<ProcessTaskStepUserVo> processTaskStepUserList = processTaskCrossoverMapper.getProcessTaskStepUserListByProcessTaskIdList(Collections.singletonList(currentProcessTaskStepVo.getProcessTaskId()));
+                System.out.println("processTaskStepUserList = " + JSONObject.toJSONString(processTaskStepUserList));
+            } else {
+                ProcessTaskStepUserVo processTaskStepUserVo = new ProcessTaskStepUserVo();
+                processTaskStepUserVo.setProcessTaskStepId(currentProcessTaskStepVo.getId());
+                processTaskStepUserVo.setUserType(ProcessUserType.MAJOR.getValue());// 只删除主处理人人
+                processTaskCrossoverMapper.deleteProcessTaskStepUser(processTaskStepUserVo);
+            }
+        }
         /* 当只分配到一个用户时，自动设置为处理人，不需要抢单 **/
         if (workerSet.size() == 1 && autoStart == 1) {
             for (ProcessTaskStepWorkerVo workerVo : workerSet) {
@@ -1691,6 +1709,10 @@ public abstract class ProcessStepHandlerBase implements IProcessStepHandler {
                     }
                 }
                 for (ProcessTaskStepUserVo oldUser : oldUserList) {
+                    /* 更新处理人状态 **/
+                    oldUser.setUserType(ProcessTaskStepUserType.HISTORY_MAJOR.getValue());
+                    oldUser.setStatus(ProcessTaskStepUserStatus.TRANSFERRED.getValue());
+                    processTaskCrossoverMapper.updateProcessTaskStepMajorUserUserTypeAndStatus(oldUser);
                     if (workerUserUuidList.contains(oldUser.getUserUuid())) {
                         String userName = oldUser.getUserName();
                         if (StringUtils.isBlank(userName)) {
@@ -1703,10 +1725,10 @@ public abstract class ProcessStepHandlerBase implements IProcessStepHandler {
                     }
                 }
                 /* 清空user表 **/
-                ProcessTaskStepUserVo processTaskStepUserVo = new ProcessTaskStepUserVo();
-                processTaskStepUserVo.setProcessTaskStepId(currentProcessTaskStepVo.getId());
-                processTaskStepUserVo.setUserType(ProcessUserType.MAJOR.getValue());
-                processTaskCrossoverMapper.deleteProcessTaskStepUser(processTaskStepUserVo);
+//                ProcessTaskStepUserVo processTaskStepUserVo = new ProcessTaskStepUserVo();
+//                processTaskStepUserVo.setProcessTaskStepId(currentProcessTaskStepVo.getId());
+//                processTaskStepUserVo.setUserType(ProcessUserType.MAJOR.getValue());
+//                processTaskCrossoverMapper.deleteProcessTaskStepUser(processTaskStepUserVo);
             }
 
             /* 保存描述内容 **/
